requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
coverage==5.3
defusedxml==0.6.0
dj-rest-auth==2.1.2
Django==3.1.4
django-allauth==0.42.0
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
Markdown==3.3.3
oauthlib==3.1.0
Pillow==8.0.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.2
